Push RuleCollection into own module for clarity
