+++ edsl/surveys/__init__.py
@@ -1,2 +1,3 @@
 from edsl.surveys.Survey import Survey
-from edsl.surveys.Rule import Rule, RuleCollection
+from edsl.surveys.Rule import Rule
+from edsl.surveys.RuleCollection import RuleCollection
