+++ tests/surveys/test_RuleCollection.py
@@ -1,6 +1,7 @@
 import unittest
 from edsl.exceptions import SurveyRuleCollectionHasNoRulesAtNodeError
-from edsl.surveys.Rule import Rule, RuleCollection
+from edsl.surveys.Rule import Rule
+from edsl.surveys.RuleCollection import RuleCollection
 
 
 class TestRuleCollection(unittest.TestCase):
